Forward `HttpRequest`s in `UpstreamHttpService` instead of `Upstream` (#116)
Make Upstream as a kind of configuration class so that forward HttpRequests in UpstreamHttpService instead of Upstream.
Remove forwarding logic from Upstream. It will manage settings of the upstream. Forwarding logic is moved to the UpstreamHttpService, but it's simple. Specific implementations such as exception mapping will be provided using SPI.
This is the last step of migration in #112.