Create spring-boot auto-configuration module (#131)
The Spring Boot is one of the most widely used framework in the JVM ecosystem. It would be convenient to provide tollgate-spring-boot2-autoconfigure and tollgate-spring-boot2-starter modules.

At first, provides a way to configure gateway server's port and protocol using properties. And to customize GatewayBuilder using beans.

And adds a simple example about how to use tollgate-spring-boot2-starter module. It will be useful for users who want to use.

Close #124