Take over remapping response status from `remapping` module to `Upstream` (#146)
Add status method to UpstreamBuilder which adds a function that remaps response status. The new remapping path function is taken over from remapping module and so deprecates the RemappingResponseStatusStrategy.

While taking over HttpStatusFunction, creates HttpStatusFunctionBuilder that helps to create HttpStatusFunction which remaps to a specified HttpStatus if a given HttpStatus is contained in specified HttpStatuses.

Regarding #146, deprecate all classes in the remapping module.