Add option to allow request or response headers (#153)
Provide a way to allow only specified headers to be included in an HTTP request to upstream or response from the upstream.

Add allowRequestHeaders and allowResponseHeaders to the UpstreamBuilder. These transform HTTP headers as an allowlist. It would be useful for those who want to hide un-intended headers from the public.

Rename DisallowRequestHeadersFunction and DisallowResponseHeadersFunction to FilteringRequestHeadersFunction and FilteringResponseHeadersFunction respectively. And allowRequestHeaders and allowResponseHeaders reuse those functions with a Predicate that returns false if contains.

Additionally, filtering doesn't remove pseudo headers.