Add mandatory properties for configuring routing (#155)
Add mandatory properties (path, upstream information such as uri or scheme and endpoints) for configuring routing on the Spring Boot integration.

It would be useful since users don't have to create GatewayConfigurator beans. The TollgateAutoConfiguration will configure a Gateway to send clients' requests to an upstream according to the Spring Boot's properties. This is an example for settings:
```
tollgate:
  gateway:
    server:
      ports:
        - port: 8080
          protocol: HTTP
    routes:
      - name: exampleProxy
        path: /
        upstream:
          uri: https://example.com
```
The Gateway will send requests for `/` to `https://example.com`.

This commit changes the properties path for the gateway server from `tollgate.server` to `tollgate.gateway.server`.